Add checks around erase in InoTable, pass useino
Pass the inode number to alloc when specified whether it is in the free
list or not.  This should be safe even in the multiple-MDS case since
specifying inodes should, in CohortFS, only be done in replication
between two already consistent local storage domains.  Put checks around
erase, so we don't run into failed assertions.