mds: create inode container directory
in MDCache::create_empty_hierarchy(), create an additional base system
inode.  this directory, called the 'inode container', will store the
primary link to all inodes in the filesystem. this will enable inodes to
be addressed directly by their inode number, and eventually remove the
need for a separate anchor table

Signed-off-by: Casey Bodley <casey@linuxbox.com>