mds: export inode with slave create message
when sending a slave create message, mknod_finish() uses
migrator->encode_export_inode() to send a copy of the inode.

slave_create_remote() uses migrator->decode_import_inode() to read the
inode and create its primary link to the inode container.  it then goes
on to clean up the inode's lock state and, in the case of mkdir, the new
directory fragment.  in the slave_create_local() case, the export state
is rolled back, and the primary link is changed from the parent dentry
to the inode container.  both remote and local cases require a lock on
the inode container dentry, so this is acquired in handle_slave_create()

handle_slave_create_ack() calls migrator->finish_export_inode() to
update the exported inode state on the originating mds, and the primary
link it created is switched to a remote link

Signed-off-by: Casey Bodley <casey@linuxbox.com>