mds: include replicas with slave_create_ack
on slave_create_ack, the master mds changes the inode's primary link to
a remote link.  this results in a cached inode with no primary parent
dentry.  the slave mds now includes replicas of the inode container
directory fragment and primary dentry with the slave_create_ack message,
allowing the master mds to cache the new inode's primary dentry.

Signed-off-by: Casey Bodley <casey@linuxbox.com>