client: readdir lists stripes in parallel
renamed struct dir_result_t to DirReader.  new struct DirChunk tracks
the entries returned by a single readdir request.  DirReader maintains
an array of DirChunks in the order they are received from the mds.  it
uses this array ordering to generate the offsets for telldir/seekdir.
this means we can only support seeks to offsets that we've encountered
within the context of the current opendir.  it also means that we can't
store the offset with the Dentry itself, because it could differ between
traversals.

new struct DirStripeReader tracks the progress of each directory stripe,
and stores the arguments passed to the next mds readdir request.

new class ReaddirWorker creates a separate thread to send and receive
readdir requests asynchronously, and will do extra readahead on a stripe
depending on the config variable mds_readdir_readahead (default 1).