Fix global CDS initialization.
CDS_Env must be instantiated before any threads are spawned, but
that can't be done until the config file and program arguments have
been parsed (since we might need to parameterize CDS).

For this reason, move CephContext::_log->start() into an init()
method called when threads are safe to spawn.

Signed-off-by: Matt Benjamin <matt@linuxbox.com>