Revert "ENOENT is not okay."
This reverts commit 3227775e555d7a6eac121e2f23fee7283460744b.

ENOENT is 'okay' if by 'okay' you mean 'does not cause the OSD to
assert'. Ceph, I am disappointed in you.

Ceph's architecture seems to check whether something ought to succeed
before the fact, and then require it to. This seems rather suboptimal to
me. Later on we when we bring in server-side erasure coding/operations
we should have a way for a multi-OSD transaction to handle failure on
one OSD in a sane fashion, since the whole point is Error Recovery.

(I suspect Ceph's design of 'crash the server' works for them since it
assumes someone else will pick up the slack later on.)