log: PrebufferedStreambuf improvements
addresses comment in Log::_flush() // FIXME: this is slow

allow PrebufferedStreambuf::get_str() to return a const ref instead of
copying the string.  this is accomplished by using a std::string only
for the internal representation, instead of having both a static buffer
and a dynamic string for overflow.

this internal std::string is sized on PrebufferedStreambuf's
constructor, which does unfortunately mean that we can't pass in an
existing buffer and use PrebufferedStreambuf to read from it.

Signed-off-by: Casey Bodley <casey@linuxbox.com>