Restore SimpleDispatcher Timings.
The simple_client/simple_server timings are based on a ping/pong
of messages between the client and server, unlike those of the
xio_client/server programs, which are one-way (so their corresponding
1-way bandwidth is appx. 2x what the test reports).  We assert
that the results are in general comparable, because in both setups,
a fixed number of messages (def. 50) is maintained in flight.

Signed-off-by: Matt Benjamin <matt@linuxbox.com>