Try-implement conditional clone of volatile buffers.
A new "strong" claim concept is added to buffer::list and buffer::ptr.
A strong claim is almost equivalent to claim, but deep copies volatile
buffers to new raw_char buffers.

A volatile buffer is one which should not be held/claimed for long
periods due to external dependencies. For example, Accelio  message
buffers are "weakly volatile."  It is safe to hold one for as long as
needed, but each is registered, and allocated from a dedicated arena.

If a caller knows that a message buffer is likely to be held for a
long period by a call path, and in particular if the buffer is
relatively small, it may be more efficient to strong claim the buffer/
buffer list before sharing it.

Strong claim is never performed unless explicitly requested.

Signed-off-by: Matt Benjamin <matt@linuxbox.com>

Conflicts:
	src/common/buffer.cc