This is a *temporary* fix to the vol/osdmap_epoch stall.
This is approach #2(subset) -- "zap global waiting_for_osdmap";
without actually zapping that list.  This implements the plumbing
to propagate osdmap changes to the vol code & to "take_waiters"
as needed.  It's a subset because I didn't prune any unused/unnecessary
code or data.

Adam plans on implementing approach #1 ("zap per-vol waiting_for_osdmap")
and that is the code you should want instead of this.  But, for now,
this gets the mds branch past that hump.

Signed-off-by: Marcus Watts <mdw@cohortfs.com>