Default volume creation
This lets you more easily configure volums by only specfying what you're
likely to care about.

This changes volume creation syntax.

But in the long term we should be able to specify key/value params,
since on the lower level MMonCommand is a parameter map, not a linear
string.

And we might want an MVolumeOp rather than relying on MMonCommand all
the time.