MDCache::_recovered() - treat "ENOENT" same as 0.
When stat'ing an object, treat "ENOENT" same as if it were empty.
This happens with 0 length files (and perhaps with other sufficiently
short files such that all the data is the inode.)  Vanilla ceph
ignores this too, but at a lower layer somewhere in the striper/filer
world which we no longer use.

Signed-off-by: Marcus Watts <mdw@cohortfs.com>