Pass r from C_Inode_Fetched::finish() - then ignore it.
It should be possible to do something intelligent with I/O errors during
inode fetch.  This implements the first part of such a fix; it passes
"r" from C_Inode_Fetched::finish to CInode::_fetched(), and logs it.

However, it's not yet possible to actually do anything with the error.
Mostly, it consists of "-ENODATA" from calls to fetch the "inode" xattr,
probably a backwards hack we should get rid of.  Until then,
ignore the error!

Signed-off-by: Marcus Watts <mdw@cohortfs.com>