Erasure coding and new structure
Turn ObjectOperation into an abstract base class (ObjOp) that is
specialized by each volume. Implement erasure coding and decoding at
this layer.

This gives OSD operations, but NOT Linger operations. This mostly means
that there is no support for Watch. This will be added later. (It's not
clear that the kind of callbacks we want match the Watch/Notify that
Ceph provides, so I don't want to let it delay our erasure any further.)