More robust error handling
Since we don't initialize the volume on decode, some operations that
always succeeded might now fail. To get around this, I have an attach
method that does the initialization, but if it's not called some
operations may fail and the result has to be checked. This is a
compromise to avoid having to spend a week peppering error checks
everywhere.

We also have a detach method for the Monitor, since that allows it to
initialize a volume (and make sure it doesn't error) and then free the
resources used by initialization. Detach is not generally thread safe,
however.