Implement shunique_lock
This is a class in the style of unique_lock and shared_lock that
implements both shared and unique locking. This is intended to fix an
infelicity where some call-chains in Objecter ended up taking both a
shared_lock and a unique_lock and taking ownership from one to the other
and vice versa.