More HAXX (test_objectstore).
The worst damage is reducing max_objects, and I did this only
to shorten the tests--the current LRU SHOULD handle arbitrary
numbers.

There is a tone of junk printing.  The printing shows that the
Synthetic test fails (apparently) only in having some incorrect
stat results, which appear to fail because opened filesystem
objects have state (did we forget to delete them, or #if this
out in the collection ops section?).

Secondarily, but nasty, the wait_till_done() logic breaks in
this case, apparently due to questionable ASSERT_* behavior--
if we allow an assertion to fail before starting the cleanup
block in SyntheticWorkload::stat(), the count of in-flight ops
will not be updated, and the test cycle will hang.  I moved
it.

Signed-off-by: Matt Benjamin <matt@cohortfs.com>

Conflicts:
	src/test/objectstore/store_test.cc