Replace std::list<Context*> with std::vector.
A std::vector of pointers is preferable to the equivalent
std::list for the usage pattern nearly everywhere in the OSD
and MDS code, mostly by avoiding allocations.  (It's cache
behavior should be better, too.)

In one special case (Paxos), I used the fact that PAXOS
proposals are a discrete subclass of Context, and used an
intrusive list, since dequeue at front is required.

Signed-off-by: Matt Benjamin <matt@cohortfs.com>

Conflicts:
	src/mds/MDCache.cc
	src/mds/MDSTable.h
	src/mds/journal.cc
	src/os/ObjectStore.h