Rework get_metadata_volume logic - start
The idea here is instead of doing lookups on the metadata volume
in an ad hoc basis all over the place, do it once near the start
of things and then pass references around as necessary.

Eventually (with multivol) the top of the mds needs to be
reorganized so that where get_metadata_volume happens today,
some sort of per-volume logic happens instead.

But right now - baby steps.  And this step isn't complete, so it it
doesn't build.

Signed-off-by: Marcus Watts <mdw@cohortfs.com>