Setup/pass volumeref through much of the code.
Since we no longer have file_layout, we can't lookup volumes all over
the place.  This is good.  But, we still need to know what volume.
So this change passes volumeref's all over the place in the client,
so they're available when needed.  Particularly noteworthy places that
now keep volume refs, the Inode and MetaRequest structures.

Signed-off-by: Marcus Watts <mdw@cohortfs.com>