Add a finisher thread, and use it on restarts.
Objecter's wait_for_map() call invokes its finisher when it has
locks outstanding.  This is unfortunate because usually one wants the
osdmap in order to issue osdmap operations, which want to do locking.
This change moves the mds operations that depend on a fresh osdmap into
a separate finisher thread that no longer has to dance around the locks,
and uses it during restarts to avoid bailing due to lock problems.

Signed-off-by: Marcus Watts <mdw@cohortfs.com>