wait_for_readable().  Don't panic if it's already readable.
The journaler has another call is_readable() that looks to to see
if it's readable (_is_readable()), and also calls prefetch().
Prefetch can make it readable before wait_for_readable checks to
see if it should wait.  Ok, we shouldn't wait, but panic'ing isn't
right.  This fix doesn't wait.  Sadly, it dies moments after this
fix, because there *is* a locking issue which is not fixed here.

Signed-off-by: Marcus Watts <mdw@cohortfs.com>