Put timestamp format back the way it was.
ctime() produces formatted times that include the spelled out month,
which are oddly enough not in alphabetical order, and the weekday,
which is not not usually useful for debugging purposes.  It omits
the # of microseconds, which I do find useful in correlating log
records between components.  ctime also appends a newline which makes
it harder to grep for things.

This diff uses strftime() to produce timestamps similar to the old
code.  It's in local time (perhaps it should be in gmtime instead.)

To be truely portable should maybe also call tzset() somewhere - I
didn't do that.

Signed-off-by: Marcus Watts <mdw@cohortfs.com>