no more data volume - so remove them from the syntax.
The mds never used the data volume - somewhere in the progression from
data pools to volumes that went away -- and we like it that way.  The
last big change got rid of it from the mds map, but failed to remove it
from the command syntax.  This makes it go away for real.

Signed-off-by: Marcus Watts <mdw@cohortfs.com>