OSD:  TLS Volume Cache
This change adds a thread-local cache of volumes, along
the same lines as the ObjectContextCache.

Ok, it's exactly the same, and I would have unified them
with a template, but didn't want to pay for the overhead
of an indirection to do the key comparisons and extra
indirection on Collection in the Object case.

Signed-off-by: Matt Benjamin <matt@cohortfs.com>