Add libosd_bench_async2
The original libosd_bench_async gets high iops, and those iops
are "real," but not representative of real workloads because
the all enqueued ops after the initial pre-staged loading are
performed by dequeue threads executing the completions.  It DOES
prove that we have at least 1.3M available iops in a "straight
through" (no thread handoff) configuration, worst case.

The new benchmark proves that we don't need to:  the new benchmark
enqueues only in the desired top-level producer thread, in batches
of the chosen queue depth.  Rather than requeue in the callback
thread, this version signals the original producer thread to
refill when each previous batch is retired (atomic counter).  The
current code, that is, doesn't even attempt to keep overlap, and
yet it still achieves over 1.3M iops--i.e., it's actually faster
than the original async version.  For this reason, and because
there are still optimizations available, it appears that we can
reliably expect over 1.5M iops on an i7 PC like mine, with only
a bit more work.

Signed-off-by: Matt Benjamin <matt@cohortfs.com>