Return type of timer functions should be void
Well, not really SHOULD be. But std::function requires return type match
and doesn't do what you would expect and swallow the return type. This
isn't necessarily bad, one could argue that the user should be made
aware that the return value is always lost, but in practice nobody
expects to have callback functions return a value and wrapping is
cumbersome.

Signed-off-by: Adam C. Emerson <aemerson@cohortfs.com>