Attached volumes and placers in a principled way
This breaks the volume and placer interfaces in two, one handling
description (and stored in the OSDMap) the other handling use.
This makes the system less flaky and error prone and lays a better
foundation for lifecycle management. It also means that we can set up
tighter integration between volumes and the objecter or OSD if we wish
without having such details leak out all over.