zfstore: use lzfw_listxattrs2 to implement rmattrs()
Use the callback from lzfw_listxattrs2 to remove all xattrs as
they are enumerated.

This hasn't been tested.  First, there is a traversal of the
listed xattr dir in progress--should we have a conditional restart
(e.g., based on the return from the cb_func)?  Is it ok to call
a function which re-enters ZFS_ENTER()?

Signed-off-by: Matt Benjamin <matt@cohortfs.com>