zfstore:  try-implement clone and clone_range
The clone_range implementation reads in 16-iov chunks, but
write 1-wise, as a simplification.  It will be easy to write
up-to the 16 chunk limit, but the extra step may need an
extra iovec array, and early exit on short read would need to
be deferred.

Fixed an apparent bug in ::read(), which didn't advance its source
offset at the 16-iov boundary.

Signed-off-by: Matt Benjamin <matt@cohortfs.com>