Fix FunQueue behavior
Due to partial update, FunQueue was accidentally peeking inside the
private function implementation of the C++ standard library. In the
libc++ case this worked by accident, in libstdc++, it did not. Now
FunQueue keeps its hands to itself.  Also changed the 'waiter' type to
be a cohort::function rather than a std::function.

This seems to fix Casey's monitor crash.

Signed-off-by: Adam C. Emerson <aemerson@linuxbox.com>